Finally finished my neovim configuration (for now).
I've been working on developing my own plugin manager. Here's how it
works:

1. I have two files in the `lua/SingularisArt/plugin` folder:
  a. `load.lua`
  b. `lazy.lua`

The first file `load.lua`, loads the given plugin. Here are some examples:

```lua
load({
  plugin = "nvim-cmp",
  config = "cmp",
})

load({
  plugin = "symbols-outline.nvim",
})

load({
  plugin = "filetype.nvim",
  config = function()
    vim.g.did_load_filetypes = 1
  end,
})
```

The first example loads the plugin `nvim-cmp`. The `config` key is the
name of the config file for the package. If you don't pass that key,
then the default file name will be the plugin name. All the
configurations are stored in `lua/SingularisArt/config/`.

The second example loads the plugin `symbols-outline.nvim`. Here, now
`config` key is used, so the plugin manager will look for the file
`lua/SingularisArt/config/symbols-outline.nvim.lua`.

The third example loads the plugin `filetype.nvim`, and a function is
being passed to the `config` key instead of the name of the config file.
This is useful when you only need to run a few lines of configuration,
instead of creating a new config file for it.

The second file `lazy.lua`, lazy loads the given plugin. Here are some examples:

```lua
lazy({
  plugin = "project.nvim",
  config = "project",
})

lazy({
  plugin = "playground",
})

lazy({
  plugin = "nvim-dap",
  config = "dap",
  event = "BufWinEnter",
  group = plugin_group,
})

lazy({
  plugin = "sniprun",
  config = "snip-run",
  event = "FileType",
  pattern = "ruby",
  group = plugin_group,
})
```

The first example loads the plugin `project.nvim`. The `config` key is the
name of the config file for the package. If you don't pass that key,
then the default file name will be the plugin name. All the
configurations are stored in `lua/SingularisArt/config/`.

The second example loads the plugin `sniprun`. As usual, the `config`
key is being passed with the name of the config file for the plugin. But
this time, there're more keys being used. The first one is the `event`.
The `lazy` function can load the plugin based on events using autocmds.
The `pattern` is the pattern, in which to be ran in. For this particular
example, the filetype to be ran in is a ruby (rb) file. The default is
`*`, meaning everything. Finally, the `group` key is the autocmd groug.
If this isn't there, then we'll have all these autocmds running with no
group to control, kill, and maintain the autocmds. You can create a
group and pass it.

Because of this, my neovim config loads in about 150 milliseconds.
The fastest time my neovim config loaded was 137 milliseconds. The longest was
189 milliseconds.

Co-Authored By: Hashem A. Damrah <singularisartt@gmail.com>